payhere/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bleach==6.0.0
Django==4.1.6
django-appconf==1.0.5
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.11
django-filter==23.1
django-imagekit==4.1.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
pilkit==2.0
Pillow==9.4.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
Unidecode==1.3.6
webencodings==0.5.1
